requirements (1).txt
opencv-python-headless
streamlit
tensorflow
keras
Pillow
numpy
cvlib
